+++ R/pase.R
@@ -1,0 +1,32 @@
+#' Data frame with sample data of PASE score questionnaire
+#'
+#' Contains non-identifiable organic trial data.
+#' Sample data labels are in Danish.
+#'
+#' @format A data frame with 200 rows and 21 variables:
+#' \describe{
+#'   \item{sample_pase01}{item 01, factor}
+#'   \item{sample_pase01b}{item 01b, factor}
+#'   \item{sample_pase02}{item 02, factor}
+#'   \item{sample_pase02a}{item 02a, factor}
+#'   \item{sample_pase03}{item 03, factor}
+#'   \item{sample_pase03b}{item 03b, factor}
+#'   \item{sample_pase04}{item 04, factor}
+#'   \item{sample_pase04b}{item 04b, factor}
+#'   \item{sample_pase05}{item 05, factor}
+#'   \item{sample_pase05b}{item 05b, factor}
+#'   \item{sample_pase06}{item 06, factor}
+#'   \item{sample_pase06b}{item 06b, factor}
+#'   \item{sample_pase07}{item 07, factor}
+#'   \item{sample_pase08}{item 08, factor}
+#'   \item{sample_pase09a}{item 09a, factor}
+#'   \item{sample_pase09b}{item 09b, factor}
+#'   \item{sample_pase09c}{item 09c, factor}
+#'   \item{sample_pase09d}{item 09d, factor}
+#'   \item{sample_pase10}{item 10, factor}
+#'   \item{sample_pase10a}{item 10a, numeric}
+#'   \item{sample_pase10b}{item 10b, factor}
+#' }
+#' @usage data(pase)
+"pase"
+